extract int reducers
